rename follow to joinStrings
