+++ 05_joinStrings/README.md
@@ -1,4 +1,4 @@
-# Exercise 05 - follow
+# Exercise 05 - joinStrings
 
 ## Step 1: Learn
 First, look at the example code in `follow-example.js`. Try and predict what the `console.log` output will be. When you're ready, run the test:
