+++ 05_joinStrings/joinStrings-example.js
