+++ 05_joinStrings/joinStrings.js
