+++ 05_joinStrings/solution/joinStrings-solution.js
