+++ 05_joinStrings/solution/joinStrings-solution.spec.js
@@ -1,25 +1,25 @@
-const values = require('./follow')
+const values = require('./joinStrings-solution')
 
 describe('follow', () => {
   test('firstName is Carlos', () => {
     expect(values.firstName).toEqual('Carlos');
   });
-  test.skip('lastName is Stevenson', () => {
+  test('lastName is Stevenson', () => {
     expect(values.lastName).toEqual('Stevenson');
   });
-  test.skip('thisYear is 1965', () => {
+  test('thisYear is 1965', () => {
     expect(values.thisYear).toEqual(1965);
   });
-  test.skip('birthYear is 1947', () => {
+  test('birthYear is 1947', () => {
     expect(values.birthYear).toEqual(1947);
   });
-  test.skip('greeting is properly output', () => {
+  test('greeting is properly output', () => {
     expect(values.greeting).toEqual('Hello! My name is Carlos Stevenson and I am 18 years old.');
   });
-  test.skip('fullName is Carlos Stevenson', () => {
+  test('fullName is Carlos Stevenson', () => {
     expect(values.fullName).toEqual('Carlos Stevenson');
   });
-  test.skip('age is 18', () => {
+  test('age is 18', () => {
     expect(values.age).toEqual(18);
   });
 });
